This element is obsolete
The HTML <input type="datetime"> was a control for entering a date and
time (hour, minute, second, and fraction of a second) as well as a
timezone. This feature has been removed from WHATWG HTML, and is no
longer supported in browsers.

Instead, browsers are implementing (and developers are encouraged to
use) <input type="datetime-local"> instead.